requirements.txt
langchain
langchain-core
langchain-community
langchain-huggingface
langchain-groq
langgraph
langgraph-checkpoint-aws
faiss-cpu
sentence-transformers
python-dotenv
typing-extensions
bedrock-agentcore